pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bus-kiosk-backend"
version = "1.0.0"
description = "Industrial Django backend for bus kiosk face recognition system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Imperial Bus Kiosk Team", email = "team@imperial-easypool.com"},
]
maintainers = [
    {name = "Imperial Bus Kiosk Team", email = "team@imperial-easypool.com"},
]
keywords = ["django", "bus", "kiosk", "face-recognition", "industrial"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    # Django Core
    "Django>=5.0,<6.0",
    "djangorestframework>=3.14.0,<4.0",
    "djangorestframework-simplejwt>=5.3.0,<6.0",

    # Database & Infrastructure
    "psycopg2-binary>=2.9.0,<3.0",
    "celery>=5.3.0,<6.0",
    "redis>=5.0.0,<6.0",
    "django-redis>=5.4.0,<6.0",

    # Monitoring & Observability
    "django-prometheus>=2.3.1,<3.0",
    "drf-api-logger>=1.1.12,<2.0",

    # Security & Performance
    "django-cors-headers>=4.3.0,<5.0",
    "django-celery-beat>=2.5.0,<3.0",
    "django-celery-results>=2.6.0,<3.0",
    "django-filter>=23.5,<24.0",
    # Removed django-ratelimit - DRF throttling is sufficient
    "django-csp>=3.8,<4.0",  # Content Security Policy

    # API Enhancement
    "djangorestframework-api-key>=2.3.0,<3.0",  # Better API key management
    # Removed drf-flex-fields - overkill for our use case
    # Removed drf-nested-routers - not needed yet

    # Utilities
    "python-decouple>=3.8,<4.0",
    "requests>=2.31.0,<3.0",
    "ulid-py>=1.1.0,<2.0",
    "drf-spectacular>=0.27.0,<1.0",
    "python-json-logger>=2.0.7,<3.0",
    "django-auditlog>=3.0.0,<4.0",  # Comprehensive audit logging
    "django-model-utils>=4.5.0,<5.0",  # Model utilities
]

[tool.setuptools]
# Explicitly define packages to avoid flat-layout issues
packages = ["bus_kiosk_backend", "users", "students", "buses", "events", "kiosks", "imperial_governance"]

[project.optional-dependencies]
# Development dependencies
dev = [
    # Type checking
    "django-stubs>=5.0.0,<6.0",
    "djangorestframework-stubs>=3.14.0,<4.0",
    "types-PyYAML>=6.0.0,<7.0",
    "types-requests>=2.31.0,<3.0",
    "types-redis>=4.6.0,<5.0",

    # Linting & Formatting (Industrial Django Stack)
    "ruff>=0.6.0,<1.0",
    "mypy>=1.8.0,<2.0",
    "black>=24.0.0,<25.0",
    "pre-commit>=3.0.0,<4.0",

    # Development tools
    "django-debug-toolbar>=4.3.0,<5.0",  # Debug toolbar
    "django-extensions>=3.2.3,<4.0",  # Management commands
    "werkzeug>=3.0.0,<4.0",  # Better debugging
]

# Testing dependencies
testing = [
    "pytest>=7.0.0,<8.0",
    "pytest-django>=4.5.0,<5.0",
    "pytest-cov>=4.0.0,<5.0",
    "factory-boy>=3.3.0,<4.0",
    "faker>=24.0.0,<25.0",  # Better test data generation
    "pytest-xdist>=3.5.0,<4.0",  # Parallel test execution
    "pytest-mock>=3.12.0,<4.0",  # Better mocking
]

# Documentation
docs = [
    "sphinx>=7.0.0,<8.0",
    "sphinx-rtd-theme>=1.3.0,<2.0",
]

[project.urls]
Homepage = "https://github.com/easypoolsol-hash/bus_kiosk_imperial"
Repository = "https://github.com/easypoolsol-hash/bus_kiosk_imperial.git"
Issues = "https://github.com/easypoolsol-hash/bus_kiosk_imperial/issues"

[tool.ruff]
# Industrial Django configuration - replaces flake8, isort, and more
target-version = "py310"
line-length = 88

[tool.ruff.lint]
# Core Python rules (comprehensive but fast)
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "B",   # flake8-bugbear (security issues)
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade (syntax upgrades)
    "DJ",  # flake8-django (Django-specific)
    "RUF", # ruff-specific rules
    "I",   # isort (import sorting)
]

ignore = [
    # Django-specific ignores (industrial best practices)
    "DJ001", # null=True on string fields (sometimes needed)
    "DJ008", # __str__ method (not always required)
    "DJ012", # Meta field order (style preference)
    "RUF012", # Mutable class attributes (Django model Meta/indexes/choices)
    # Django patterns that are acceptable
    "BLE001",  # blind except (Django uses this pattern)
    "C901",    # too complex (Django models can be complex)
    "N805",    # self naming (Django conventions)
]

[tool.ruff.lint.per-file-ignores]
# Django-specific file rules
"*/migrations/*.py" = ["E501", "F401"]  # Migrations can be long
"*/settings/*.py" = ["F401"]  # Settings import patterns
"*/tests/*.py" = ["B011", "S101"]  # Test patterns
"**/models.py" = ["RUF012"]  # Django model class attributes (Meta, choices, etc.)
"manage.py" = ["T201"]  # Django management script

[tool.ruff.lint.isort]
# Django-aware import sorting
known-first-party = ["bus_kiosk_backend", "users", "students", "buses", "events", "kiosks"]
force-sort-within-sections = true

[tool.black]
# Code formatting (complements ruff)
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs|\.git|\.hg|\.mypy_cache|\.tox|\.venv|build|dist|migrations
)/
'''
